requirements.txt
python-dotenv==1.0.0
black==23.11.0